refactor(product): preserve complexity budgets
