+++ framework/core/src/python/kungfu/cli/commands/assignment.py
@@ -2097,39 +2097,6 @@ def _identity_options(function):
     return function
 
 
-def _exact_pending_fit_review(
-    current,
-    *,
-    missing_message,
-    conflicting_message,
-):
-    reviews = list(current.get("independent_reviews") or [])
-    decisions = list(current.get("continuation_decisions") or [])
-    decided_review_ids = {str(row.get("review_id") or "") for row in decisions}
-    pending = [
-        row
-        for row in reviews
-        if str(row.get("review_id") or "") not in decided_review_ids
-        and row.get("verdict") == "fit"
-    ]
-    if not pending:
-        raise ValueError(missing_message)
-    exact_fit_roots = {
-        (
-            str(row.get("claim_id") or ""),
-            str(row.get("claim_payload_hash") or ""),
-            str(row.get("continuation_plan_root") or ""),
-        )
-        for row in pending
-    }
-    if len(exact_fit_roots) != 1:
-        raise ValueError(conflicting_message)
-    # Repeated product requests can retain more than one independent review
-    # event for the same exact completion claim and continuation plan. They
-    # are equivalent evidence, not an ambiguous human decision.
-    return pending[-1]
-
-
 def _work_close_plan(
     *,
     workspace_root,
@@ -2144,7 +2111,7 @@ def _work_close_plan(
     decision_mode = "required"
     decision = None
     if current["phase"] == "independently-reviewed":
-        review = _exact_pending_fit_review(
+        review = assignment_review.exact_pending_fit_review(
             current,
             missing_message=(
                 "Work close requires one exact undecided fit independent review"
@@ -2256,7 +2223,7 @@ def _resume_starter_close(
     review_receipt = None
     close_receipt = None
     if current["phase"] == "independently-reviewed":
-        review = _exact_pending_fit_review(
+        review = assignment_review.exact_pending_fit_review(
             current,
             missing_message="retained review state is missing one exact fit review",
             conflicting_message="retained review state has conflicting fit reviews",
