+++ shifu
@@ -306,23 +306,13 @@ if [ "${1:-}" = "work" ]; then
   printf '%s\n' '{"schema":"kungfu.assignment-orchestration.diagnosis/v1","ok":false,"code":"assignment-current-checkout-binding-missing","message":"Assignment admission requires pykungfu from the current checkout","next_actions":[{"action":"build-core","command":"./shifu build:core","description":"Assemble pykungfu from the current checkout"}]}'
   exit 127
 fi
-# The agent-facing Work Control CLI must be the executable assembled from
-# this checkout.  Bypass pnpm here so --json remains a single JSON document and
-# cannot silently fall through to an unrelated globally installed `kungfu`.
+# Use the executable assembled from this checkout, never a global `kungfu`.
 if [ "${1:-}" = "kungfu" ]; then
   shift
   _kungfu_checkout_bin="./framework/core/dist/kungfu/kungfu"
   if [ -x "$_kungfu_checkout_bin" ]; then
-    if [ -z "${KUNGFU_TUI_ENTRY:-}" ] \
-      && [ -f "./framework/tui/dist/tui.mjs" ]; then
-      KUNGFU_TUI_ENTRY="$PWD/framework/tui/dist/tui.mjs"
-      export KUNGFU_TUI_ENTRY
-    fi
-    if [ -z "${KF_BUNDLED_EXTENSION_ROOT:-}" ] \
-      && [ -f "./product/extensions/agent-work-lab/experience/starter-project.json" ]; then
-      KF_BUNDLED_EXTENSION_ROOT="$PWD/product/extensions"
-      export KF_BUNDLED_EXTENSION_ROOT
-    fi
+    [ -n "${KUNGFU_TUI_ENTRY:-}" ] || [ ! -f "./framework/tui/dist/tui.mjs" ] || export KUNGFU_TUI_ENTRY="$PWD/framework/tui/dist/tui.mjs"
+    [ -n "${KF_BUNDLED_EXTENSION_ROOT:-}" ] || [ ! -f "./product/extensions/agent-work-lab/experience/starter-project.json" ] || export KF_BUNDLED_EXTENSION_ROOT="$PWD/product/extensions"
     exec "$_kungfu_checkout_bin" "$@"
   fi
   echo "shifu: kungfu source CLI is not assembled; run ./shifu build:core" >&2
